📦 chore(docker): add Dockerfiles for AstronVim, LazyVim, LunarVim, and NvChad 🔥 refactor(module): remove neovim.nu module and replace with nvim.nu module
The Dockerfiles for AstronVim, LazyVim, LunarVim, and NvChad have been added to the respective directories in the docker folder. These Dockerfiles contain the necessary instructions to build the Docker images for each Vim distribution.

The neovim.nu module has been removed and replaced with the nvim.nu module. This change was made to align with the naming convention used in the Dockerfiles and to improve consistency across the codebase.